requirements.txt
customtkinter>=5.2.0
httpx>=0.27.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
pyinstaller>=6.0.0
